Add extra comments and more helper function.
I'm so amazed for both asami and @joinr demo to make asami durable.

Idk, but I feel I should turn this into a small library that can
be improved on further as my knowledge of understanding of data serialization
and nippy increases.

As well to easily use those helper functions in projects, with needing
to redefine it.

Many thanks to @joinr for making this possible!!